Adding new configs more favorable to method.
Adding three new configs with distributions on population size
expansions and times that are meant to be "easy" for the method to see
if there is any region of parameter space where it works well.

Our last attempt at this was too extreme. The gamma(shape=100, mean=0.1)
prior on the relative root pop size lead to MCMC problems.
Basically, when there was little info on the time of the expansion (bc
it was old relative to the pop size), the chains would often find
local optima where the time was near zero, and the root size was
close to the true size of the descendant pop size. Because the
prior on the relative sizes of the root/descendant was so strong,
it was difficult to get out of this region of parameter space.

Because such a prior would never be used in practice, it's not worth
trying to develop new MCMC proposals to solve this issue.  The new
configs will explore more reasonable settings.