Removing simulation results under extreme priors.
Removing simulation results under the extreme population expansion
settings. The prior on the expansion was too extreme and informative
such that when there was little or no info about the time of the
expansion (too old or too recent), the MCMC chain would get stuck in
local optima. Because these priors would never be used in practice, it's
not worth developing new MCMC behavior to work well in this corner case.