Adding results of diffuse prior pairs sims.
The results of analyses of simulated data with 3 pairs (divergences)
under diffuse priors was not added previously (I thought it was).  It
was missed, because the python parsing script did not visit the relevant
simulation directory. Updating script to catch this, which required a
few other tweaks, because the population labels for the divergence pairs
were different that the pop-size-change comparisons. These updates and
the simulation results parsed/summarized by the script are being added.